Moved all functions and actions to separate AudioPlayerManager that will handle music across views & for less clutter.
